requirements.txt
flask
requests
pandas
python-dotenv
groq